EMBR-7161 move types to dedicated file, type protection for one of appID or custom export
